Separated out Home screen, to make adding auth easier.
