+++ package.json
@@ -38,6 +38,9 @@
   },
   "jest": {
     "preset": "react-native",
+    "transformIgnorePatterns": [
+      "node_modules/(?!(jest-)?@?react-native|@react-native-community|@react-navigation)"
+    ],
     "moduleFileExtensions": [
       "ts",
       "tsx",
